requirements.txt
numpy~=1.21.4
pandas~=1.3.5
tqdm~=4.62.3
torch~=1.10.0
hydra-core~=1.1.1
omegaconf~=2.1.1